Develop (#8)
* chore: Add shadow to navbar for better visual appeal

The code changes in this commit add a shadow to the navbar component in order to improve its visual appeal. This change is made in the `Navbar.tsx` file, specifically in the `nav` element. The `shadow-md` class is added to the `className` attribute to achieve the desired effect.

* chore(log): delete console log in login and publication description

* feat(signup): Add birthdate validation to signup form

The code changes in this commit add birthdate validation to the signup form. It calculates the age based on the provided birthdate and displays an error message if the age is less than 18. This change is made in the `Signup.tsx` file.

* chore(publications create): Remove console logs and disable input fields in PublicationForm

The code changes in this commit remove console logs and disable input fields in the PublicationForm component. This improves the code by removing unnecessary logging statements and preventing users from editing certain fields.

* feat: add genres to book

* refactor: update landing page layout

* refactor: make text visible

* feat: fix iniciar/cerrar sesión

* feat: add email confirmation

* refactor: change publication types' definitions

* refactor: update favicon and page title

* feat: set default genres to ["N/A"] in PublicationForm

* refactor: join genres with comma in publication description

---------

Co-authored-by: pacristi <piero.cristi@uc.cl>
Co-authored-by: Piero Cristi Monasterio <62040073+pacristi@users.noreply.github.com>